chore: move pieces models to shared package
