+++ packages/shared/src/lib/pieces/dto/get-piece-request-params.ts
@@ -1,7 +1,7 @@
 import { Static, Type } from "@sinclair/typebox";
 
 export const GetPieceRequestParams = Type.Object({
-    name: Type.String({}),
+    name: Type.RegEx(/[A-Za-z0-9_]+/),
 });
 
 export type GetPieceRequestParams = Static<typeof GetPieceRequestParams>;
