+++ packages/shared/src/lib/pieces/dto/get-piece-request-query.ts
@@ -1,7 +1,7 @@
 import { Static, Type } from "@sinclair/typebox";
 
 export const GetPieceRequestQuery = Type.Object({
-    version: Type.String({}),
+    version: Type.RegEx(/[0-9]+\.[0-9]+\.[0-9]+/),
 });
 
 export type GetPieceRequestQuery = Static<typeof GetPieceRequestQuery>;
