+++ packages/shared/src/lib/pieces/index.ts
@@ -1,0 +1,5 @@
+export * from './dto/get-piece-request-params'
+export * from './dto/get-piece-request-query'
+export * from './dto/piece-option-request'
+export * from './model/piece-metadata'
+export * from './model/property-type'
